Adds position capabilities editor details to player editor command
